requirements.txt
fastapi>=0.124.0
boto3>=1.28.0
rich>=13.0.0
SQLAlchemy
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
httpx
aiosqlite
pytest
pytest-asyncio
python-dateutil